+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Changed
+
+- Single-day closures will have a single date displayed instead of the format "{{startDate}} to {{endDate}}"
+
 ## [0.2.1] - 2021-06-07
 
 ### Fixed